tcb: gitlab-ci: Fix the use of cache
TCB build runs in some steps, that depends on some files that are
generated in the previous steps. This files are not tracked by git,
so they are not cached by gitlab. This commit adds the cache policy
to the steps that generate these files, so they can be used in the
next steps.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>