Improve the compilation and packaging of dotnet release code
Improve the compilation and packaging of dotnet release code on the
Dockerfile. With "dotnet build" and "dotnet publish" the code gets
compiled twice, because "dotnet publish" also compiles the code. Also,
by default the code gets compiled to Debug, but we want it to Release,
so the "-c Release" is necessary.
Related-to: TIE-716
Signed-off-by: Andre Riesco <andre.riesco@toradex.com>