cppQML: Add the hardware acceleration support for the Qt template
Add the hardware acceleration support for the Qt template, by
using the newly release Torizon Qt6 containers (torizon/qt6-wayland and
torizon/qt6-wayland-vivante).
Also change the build tool from qmake6 to cmake, because at the moment
the "quick" module when added through qmake6 (in the .pro file),
adds the requirement of libGLX on the Makefile, which is not available
for the iMX8 modules.
Therefore, before it gets fixed, to use qmake6 you have remove the
/usr/lib/aarch64-linux-gnu/libGLX.so from the LIBS of the Makefile
generated by qmake6 before running "make". Because of that, the
recommended way is to use CMake, the default of this new
hardware accelerated Qt6 template.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>