cpp: Change SSH executables copy to Docker COPY and standardize app root
This commit:
. Changes the copy of executables from the SSH on the running
container to copy them with COPY in the Dockerfiles on the container
build.
. Standardizes between the SDK, the debug and the production the
APP_ROOT directory variable, to use a value defined on settings.json,
making it clear for the user where his project files are.
Does the same for the APP_EXECUTABLE file.
. Adds the information about how to deal with extra files (not added to
the running container automatically) on the documentation of each
template.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>