scripts: shareWSLPorts: Add debug option
This is needed because the script opens a new window in the Windows side,
and it closes immediately after the script finishes. This way we can't
see the output of the script. Adding a debug option will make the script
wait for the user to press enter before closing the window.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>