dotnetUno5FrameBuffer: add console exit handler
When the user has a keyboard attached to the device, it is possible to
exit the application by pressing F12. But in a headless environment, it
is not possible to do that. This change adds a console exit handler to
the application that allows the user to exit the application by pressing
F12 in the debug terminal attached to the application stdout.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>