scripts: projectUpdater: Fix script asking duplicate confirmation
If the script update itself is called from the projectUpdater script, it
will ask for confirmation twice, this is because the script is called
twice, the first time to update the project and the second time to
update the projectUpdater script itself. This patch adds a new argument
to the script to avoid asking for confirmation on the second run.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>