scripts: tcb: Fix Github Actions dind workdir mount
When running the tcb in a docker-in-docker environment, the workdir
mount point is not found, because the github actions runner sandbox
the runner by re-writing the absolute path to a different location.
So, first we need to get the absolute path of the workspace and then
store it in a file called abs-path, so we can use get the path from
the file and use it as the workdir mount point.

Reported-by: @dk-coligo
Related-to: #202
Signed-off-by: Matheus Castello <matheus.castello@toradex.com>