common.json: apply-torizon-packages to apply-torizon-packages-<arch>
Change apply-torizon-packages to apply-torizon-packages-<arch>, to
pass the arch instead of getting it from settings.json. This task is
only used as dependsOn of task that pass the arch instead of getting
them from settings.json. So this avoids applying arm64 packages if it
is the default device when running for example
build-container-torizon-release-arm.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>